test(animations): should the 'slideRight' animation return something
