+++ src/shared/animations/__test__/animations.spec.js
@@ -1,4 +1,4 @@
-import { slideTop, splitview, slideLeft, slideRight } from '../index.js'
+import { slideTop, slideLeft, slideRight } from '../index.js'
 
 describe('animations', () => {
   it('should the "slideTop" animation return something', () => {
@@ -7,5 +7,9 @@ describe('animations', () => {
 
   it('should the "slideLeft" animation return something', () => {
     expect(slideLeft(document.createElement("div"))).toBeDefined()
+  })
+
+  it('should the "slideRight" animation return something', () => {
+    expect(slideRight(document.createElement("div"))).toBeDefined()
   })
 })
