+++ src/shared/animationsWinJs/index.js
@@ -7,3 +7,15 @@ export default function (isAnimate) {
         WinJS.UI.disableAnimations()
     }
 }
+
+const animate = () => localStorage.getItem('display') ? JSON.parse(localStorage.getItem('display')).animations : {}
+
+const splitview = (element, expanded) => {
+    const animation = element.animate({
+        width: expanded ? ['0px', '200px'] : ['200px', '0px']
+    }, 150)
+    if (!animate()) animation.play = () => {}
+    return animation 
+}
+
+export { splitview }
